--- pkg/apply/gvk.go
@@ -1,96 +1,0 @@
-// Copyright 2020 The Kubernetes Authors.
-// SPDX-License-Identifier: Apache-2.0
-
-package apply
-
-import (
-	"strings"
-)
-
-// Gvk identifies a Kubernetes API type.
-// https://github.com/kubernetes/community/blob/master/contributors/design-proposals/api-machinery/api-group.md
-type Gvk struct {
-	Group   string `json:"group,omitempty" yaml:"group,omitempty"`
-	Version string `json:"version,omitempty" yaml:"version,omitempty"`
-	Kind    string `json:"kind,omitempty" yaml:"kind,omitempty"`
-}
-
-// Values that are brief but meaningful in logs.
-const (
-	noGroup   = "~G"
-	noVersion = "~V"
-	noKind    = "~K"
-	fieldSep  = "_"
-)
-
-// String returns a string representation of the GVK.
-func (x Gvk) String() string {
-	g := x.Group
-	if g == "" {
-		g = noGroup
-	}
-	v := x.Version
-	if v == "" {
-		v = noVersion
-	}
-	k := x.Kind
-	if k == "" {
-		k = noKind
-	}
-	return strings.Join([]string{g, v, k}, fieldSep)
-}
-
-// Equals returns true if the Gvk's have equal fields.
-func (x Gvk) Equals(o Gvk) bool {
-	return x.Group == o.Group && x.Version == o.Version && x.Kind == o.Kind
-}
-
-// An attempt to order things to help k8s, e.g.
-// a Service should come before things that refer to it.
-// Namespace should be first.
-// In some cases order just specified to provide determinism.
-var orderFirst = []string{
-	"Namespace",
-	"ResourceQuota",
-	"StorageClass",
-	"CustomResourceDefinition",
-	"MutatingWebhookConfiguration",
-	"ServiceAccount",
-	"PodSecurityPolicy",
-	"Role",
-	"ClusterRole",
-	"RoleBinding",
-	"ClusterRoleBinding",
-	"ConfigMap",
-	"Secret",
-	"Service",
-	"LimitRange",
-	"PriorityClass",
-	"Deployment",
-	"StatefulSet",
-	"CronJob",
-	"PodDisruptionBudget",
-}
-var orderLast = []string{
-	"ValidatingWebhookConfiguration",
-}
-var typeOrders = func() map[string]int {
-	m := map[string]int{}
-	for i, n := range orderFirst {
-		m[n] = -len(orderFirst) + i
-	}
-	for i, n := range orderLast {
-		m[n] = 1 + i
-	}
-	return m
-}()
-
-// IsLessThan returns true if self is less than the argument.
-func (x Gvk) IsLessThan(o Gvk) bool {
-	indexI := typeOrders[x.Kind]
-	indexJ := typeOrders[o.Kind]
-	if indexI != indexJ {
-		return indexI < indexJ
-	}
-	return x.String() < o.String()
-}
